Add automatic matching back

+++ src/server/api/routers/matchRequest.ts
@@ -2,16 +2,20 @@
 /* eslint-disable @typescript-eslint/no-unsafe-call */
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /* eslint-disable @typescript-eslint/no-unsafe-return */
+import { TRPCError } from "@trpc/server";
 import { observable } from "@trpc/server/observable";
 import { EventEmitter } from "events";
 import { z } from "zod";
 import { createTRPCRouter, protectedProcedure } from "~/server/api/trpc";
+import { prismaPostgres } from "~/server/db";
+import { MatchType } from "@prisma-db-psql/client"
 
 const userObject = z.object({
   id: z.string(),
   name: z.string(),
   difficulty: z.number().min(0).max(5),
   category: z.string(),
+  matchType: z.enum(["AUTO", "MANUAL"])
 });
 
 const addJoinRequestObject = z.object({
@@ -30,10 +34,29 @@ type UserRequest = {
 const ee = new EventEmitter();
 
 export const matchRequestRouter = createTRPCRouter({
-  getAllRequests: protectedProcedure.query(async ({ ctx }) => {
-    return await ctx.prismaPostgres.matchRequest.findMany({
-
+  getAllOtherRequests: protectedProcedure
+  .query(async ({ ctx }) => {
+    const requests = await ctx.prismaPostgres.matchRequest.findMany({
+      where: {
+        NOT: [
+          {
+            OR: [
+              {
+                userId: ctx.session.user.id
+              },
+              {
+                matchType: "AUTO"
+              }
+            ]
+          }
+        ]
+      }
     });
+    return {
+      requests,
+      count: await ctx.prismaPostgres.matchRequest.count()
+    }
+
   }),
   getOwnRequest: protectedProcedure.query(async ({ ctx }) => {
     const ownRequest = await ctx.prismaPostgres.matchRequest.findUnique({
@@ -64,7 +87,7 @@ export const matchRequestRouter = createTRPCRouter({
   addRequest: protectedProcedure
     .input(userObject)
     .mutation(async ({ ctx, input }) => {
-      const { id, name, difficulty, category } = input;
+      const { id, name, difficulty, category, matchType } = input;
 
       const existingRequest = await ctx.prismaPostgres.matchRequest
         .findFirst({
@@ -95,6 +118,7 @@ export const matchRequestRouter = createTRPCRouter({
             name,
             difficulty,
             category,
+            matchType
           },
         })
         .then((req) => {
@@ -104,7 +128,7 @@ export const matchRequestRouter = createTRPCRouter({
       ee.emit("add", request);
 
       return {
-        msg: "Searching for partner...",
+        msg: request.matchType == "MANUAL" ? "Room created" : "Searching for partner...",
         partner: "",
         isSuccess: true,
         difficulty,
@@ -171,23 +195,40 @@ export const matchRequestRouter = createTRPCRouter({
           category,
         },
       });
-
+      ee.emit("update", updatedRequest);
       return updatedRequest;
     }),
 
   subscribeToAllRequests: protectedProcedure.subscription(() => {
-    return observable<UserRequest>((emit) => {
+    return observable<UserRequest & {
+      mode: "add" | "remove" | "update"
+    }>((emit) => {
+      // todo(gab): Consider using only one since currently the subscription only tells to refetch
       const onAdd = (data: UserRequest) => {
-        emit.next(data);
+        emit.next({
+          ...data,
+          mode: "add",
+        });
       };
       const onRemove = (data: UserRequest) => {
-        emit.next(data);
+        emit.next({
+          ...data,
+          mode: "remove",
+        });
       };
+      const onUpdate = (data: UserRequest) => {
+        emit.next({
+          ...data,
+          mode: "update"
+        })
+      }
+      ee.on("update", onUpdate)
       ee.on("add", onAdd);
       ee.on("remove", onRemove);
       return () => {
         ee.off("add", onAdd);
         ee.off("remove", onRemove);
+        ee.off("update", onUpdate)
       };
     });
   }),
@@ -222,7 +263,61 @@ export const matchRequestRouter = createTRPCRouter({
         ee.emit("join", { joiningUser, joiningUserId, originalRequestId });
       }
     }),
+    notifyAutomaticRequests: protectedProcedure
+      .input(z.object({ requestId: z.string() }))
+      .mutation(({ ctx, input }) => {
+        ee.emit("findAutomatic", { requestId: input.requestId, userId: ctx.session.user.id })
+      })
+      ,
+  /**
+   * Endpoint needed to ensure that no too users are matched repeatedly
+   */
+  subscribeToAutomaticRequests: protectedProcedure  
+    .subscription(({ ctx }) => {
+      return observable<{user1Id: string, user2Id: string}>((emit) => {
+        // when a request is found, emit the two users
+        async function find({requestId, userId}: { requestId: string, userId: string}) {
+          const req = await prismaPostgres.matchRequest.findUnique({
+            where: {
+              id: requestId
+            }
+          });
+          if (!req) return;
+          const other = await prismaPostgres.matchRequest.findFirst({
+            where: {
+              category: req.category,
+              difficulty: req.difficulty,
+              matchType: "AUTO",
+              NOT: {
+                id: requestId
+              }
+            }
+          });
+          if (other) {
+            await prismaPostgres.matchRequest.deleteMany({
+              where: {
+                userId: {
+                  in: [userId, other.userId]
+                }
+              }
+            });
 
+            emit.next({
+              user1Id: ctx.session.user.id,
+              user2Id: other.id
+            });
+          }
+        }
+        const findSync = (arg1: { requestId: string, userId: string }) => {
+          console.log(arg1);
+          void find(arg1);
+        }
+        ee.on("findAutomatic", findSync);
+        return () => {
+          ee.off("findAutomatic", findSync);
+        }
+      })
+    }),
   // Users listens to other users who want to join their session
   subscribeToJoinRequests: protectedProcedure.subscription(() => {
     return observable<{
@@ -249,7 +344,20 @@ export const matchRequestRouter = createTRPCRouter({
     .input(z.object({ acceptId: z.string(), requestId: z.string() }))
     .mutation(async ({ ctx, input }) => {
       const { acceptId, requestId } = input;
-
+      const r1 = await ctx.prismaPostgres.matchRequest.findUnique({
+        where: {
+          id: acceptId,
+        }
+      });
+      const r2 = await ctx.prismaPostgres.matchRequest.findUnique({
+        where: {
+          id: requestId
+        }
+      })
+      if (!r1 || !r2) throw new TRPCError({
+        code: "BAD_REQUEST", 
+        message: "Failed to match requests"
+      })
       await ctx.prismaPostgres.matchRequest.deleteMany({
         where: {
           id: { in: [acceptId, requestId] },
@@ -262,8 +370,13 @@ export const matchRequestRouter = createTRPCRouter({
           toId: { in: [acceptId, requestId] },
         },
       });
-
+      
       ee.emit("confirm", { acceptId, requestId });
+      // needed for matching information
+      return {
+        user1Id: r1.userId,
+        user2Id: r2.userId
+      }
     }),
 
   // Users listens for confirmation from other user to join the session
